refactor phase 2
